applications/backend/services/clicks/src/pyproject.toml
[tool.poetry]
name = "Clicker"
version = "0.0.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.109.0"
uvicorn = "^0.25.0"
codefly-sdk = "^0.0.10"
psycopg2-binary = "^2.9.9"
sqlalchemy = "^2.0.29"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
